strip down unidoc-build to bare essentials, deprecate need to specify 'local' subcommand
