+++ Sources/UnidocClient/Unidoc.BuildArtifact (ext).swift
