+++ Sources/UnidocClient/exports.swift
@@ -1,0 +1,2 @@
+@_exported import enum SymbolGraphCompiler.SSGC
+@_exported import enum Unidoc.Unidoc
