+++ Sources/unidoc-build/Unidoc.Build.swift
@@ -1,22 +1,14 @@
-#if canImport(Glibc)
-@preconcurrency import Glibc
-#elseif canImport(Darwin)
-@preconcurrency import Darwin
-#endif
-
 import ArgumentParsing
 import HTTP
+import SymbolGraphCompiler
 import SymbolGraphs
 import Symbols
 import System
-import Unidoc
-import UnidocAPI
 
 extension Unidoc
 {
     struct Build:Sendable
     {
-        var authorization:String?
         var project:Symbol.Package?
         var host:String
         var port:Int
@@ -30,13 +22,11 @@ extension Unidoc
         var swiftRuntime:String?
         var swiftPath:String?
         var swiftSDK:SSGC.AppleSDK?
-        var force:Unidoc.VersionSeries?
         var input:String?
 
         private
         init()
         {
-            self.authorization = nil
             self.project = nil
             self.host = "localhost"
             self.port = 8080
@@ -48,7 +38,6 @@ extension Unidoc
             self.swiftRuntime = nil
             self.swiftPath = nil
             self.swiftSDK = nil
-            self.force = nil
             self.input = nil
         }
     }
@@ -60,8 +49,6 @@ extension Unidoc.Build
     static
     func main() async
     {
-        setlinebuf(stdout)
-
         var arguments:CommandLine.Arguments = .init()
         guard
         let command:String = arguments.next()
@@ -79,22 +66,7 @@ extension Unidoc.Build
         do
         {
             let build:Self = try .parse(arguments: arguments)
-
-            switch command
-            {
-            case "requests":
-                try await build.requests()
-
-            case "local":
-                try await build.local()
-
-            case "latest":
-                try await build.latest()
-
-            case let command:
-                print("Unknown command: \(command)")
-                SystemProcess.exit(with: 1)
-            }
+            try await build.local()
         }
         catch let error
         {
@@ -114,16 +86,6 @@ extension Unidoc.Build
         {
             switch option
             {
-            case "--authorization", "-i":
-                options.authorization = try arguments.next(for: option)
-                options.port = 8443
-
-            case "--swiftinit", "-S":
-                options.host = "swiftinit.org"
-                options.port = 443
-
-                options.authorization = try arguments.next(for: option)
-
             case "--host", "-h":
                 options.host = try arguments.next(for: option)
 
@@ -148,11 +110,8 @@ extension Unidoc.Build
             case "--input", "-I":
                 options.input = try arguments.next(for: option)
 
-            case "--force", "-f":
-                options.force = .release
-
-            case "--force-prerelease", "-e":
-                options.force = .prerelease
+            case "local":
+                print("Warning: the 'local' subcommand is deprecated and no longer necessary.")
 
             case let option:
                 if  case nil = options.project
@@ -185,55 +144,6 @@ extension Unidoc.Build
 extension Unidoc.Build
 {
     private
-    func requests() async throws
-    {
-        let unidoc:Unidoc.Client<HTTP.Client2> = try .init(from: self)
-        let cache:FilePath = "swiftpm"
-
-        while true
-        {
-            //  Don’t run too hot if the network is down.
-            async
-            let cooldown:Void = try await Task.sleep(for: .seconds(5))
-
-            do
-            {
-                let labels:Unidoc.BuildLabels? = try await unidoc.connect
-                {
-                    try await $0.labels()
-                }
-
-                if  let labels:Unidoc.BuildLabels
-                {
-                    print("""
-                        Building package '\(labels.package)' at '\(labels.ref)' \
-                        (\(labels.coordinate))
-                        """)
-
-                    /// As this runs continuously, we should remove the build artifacts
-                    /// afterwards, to avoid filling up the disk. We must also remove the cloned
-                    /// repository, as it may experience name conflicts on long timescales.
-                    try await unidoc.buildAndUpload(
-                        labels: labels,
-                        action: .uplinkRefresh,
-                        remove: true,
-                        cache: cache)
-                }
-                else
-                {
-                    print("Heartbeat received; no packages to build.")
-                }
-            }
-            catch let error
-            {
-                print("Error: \(error)")
-            }
-
-            try await cooldown
-        }
-    }
-
-    private
     func local() async throws
     {
         guard
@@ -246,46 +156,7 @@ extension Unidoc.Build
         let search:FilePath? = self.input.map(FilePath.init(_:))
         let type:SSGC.ProjectType = self.book ? .book : .package
 
-        if  case nil = self.authorization
-        {
-            let unidoc:Unidoc.Client<HTTP.Client1> = try .init(from: self)
-            try await unidoc.buildAndUpload(local: project, search: search, type: type)
-        }
-        else
-        {
-            let unidoc:Unidoc.Client<HTTP.Client2> = try .init(from: self)
-            try await unidoc.buildAndUpload(local: project, search: search, type: type)
-        }
-    }
-
-    private
-    func latest() async throws
-    {
-        let unidoc:Unidoc.Client<HTTP.Client2> = try .init(from: self)
-
-        guard
-        let project:Symbol.Package = self.project
-        else
-        {
-            print("No project specified!")
-            return
-        }
-
-        let labels:Unidoc.BuildLabels? = try await unidoc.connect
-        {
-            try await $0.labels(of: project, series: self.force ?? .release)
-        }
-
-        guard
-        let labels:Unidoc.BuildLabels
-        else
-        {
-            print("Package '\(project)' is not buildable by labels!")
-            return
-        }
-
-        try await unidoc.buildAndUpload(
-            labels: labels,
-            action: self.force != nil ? .uplinkRefresh : .uplinkInitial)
+        let unidoc:Unidoc.Client<HTTP.Client1> = try .init(from: self)
+        try await unidoc.buildAndUpload(local: project, search: search, type: type)
     }
 }
