+++ Sources/unidoc-build/exports.swift
@@ -1,2 +1,1 @@
-@_exported import enum SymbolGraphCompiler.SSGC
 @_exported import enum Unidoc.Unidoc
